Move stats classes to results package
